requirements.txt
aiosmtplib==1.1.6
altgraph==0.17.2
anyio==3.5.0
appdirs==1.4.4
APScheduler==3.6.3
arabic-reshaper==2.1.3
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
cryptography==35.0.0
cycler==0.10.0
dataclasses-json==0.5.6
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
Django==3.0
django-createsuperuser==2020.12.3
django-createsuperuserwithpassword==2.0.0
django-filter==21.1
django-import-export==2.7.1
django-jazzmin==2.4.9
django-social-share==2.2.1
djlint==0.7.6
docopt==0.6.2
et-xmlfile==1.1.0
feedparser==6.0.8
filelock==3.0.12
Flask==2.0.2
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
h11==0.12.0
html5lib==1.1
httpcore==0.13.7
httpx==0.19.0
idna==3.3
imageio==2.10.3
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.3
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.2
Js2Py==0.71
jsonfield==3.1.0
kiwisolver==1.3.1
macholib==1.15.2
MarkupPy==1.14
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
matplotlib==3.4.2
moviepy==1.0.3
mutagen==1.45.1
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.26
mysql-connector-python-rf==2.2.2
num2words==0.5.10
numpy==1.21.0
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.0
pathspec==0.9.0
Pillow==8.4.0
pipwin==0.5.1
proglog==0.1.9
protobuf==3.18.1
PyAudio==0.2.11
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.11.0
pycryptodomex==3.11.0
pyfiglet==0.8.post1
Pygments==2.10.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyjsparser==2.7.1
PyJWT==2.1.0
PyMySQL==1.0.2
pyobjc==7.3
pyobjc-core==7.3
pyobjc-framework-Accessibility==7.3
pyobjc-framework-Accounts==7.3
pyobjc-framework-AddressBook==7.3
pyobjc-framework-AdServices==7.3
pyobjc-framework-AdSupport==7.3
pyobjc-framework-AppleScriptKit==7.3
pyobjc-framework-AppleScriptObjC==7.3
pyobjc-framework-ApplicationServices==7.3
pyobjc-framework-AppTrackingTransparency==7.3
pyobjc-framework-AuthenticationServices==7.3
pyobjc-framework-AutomaticAssessmentConfiguration==7.3
pyobjc-framework-Automator==7.3
pyobjc-framework-AVFoundation==7.3
pyobjc-framework-AVKit==7.3
pyobjc-framework-BusinessChat==7.3
pyobjc-framework-CalendarStore==7.3
pyobjc-framework-CallKit==7.3
pyobjc-framework-CFNetwork==7.3
pyobjc-framework-ClassKit==7.3
pyobjc-framework-CloudKit==7.3
pyobjc-framework-Cocoa==7.3
pyobjc-framework-Collaboration==7.3
pyobjc-framework-ColorSync==7.3
pyobjc-framework-Contacts==7.3
pyobjc-framework-ContactsUI==7.3
pyobjc-framework-CoreAudio==7.3
pyobjc-framework-CoreAudioKit==7.3
pyobjc-framework-CoreBluetooth==7.3
pyobjc-framework-CoreData==7.3
pyobjc-framework-CoreHaptics==7.3
pyobjc-framework-CoreLocation==7.3
pyobjc-framework-CoreMedia==7.3
pyobjc-framework-CoreMediaIO==7.3
pyobjc-framework-CoreMIDI==7.3
pyobjc-framework-CoreML==7.3
pyobjc-framework-CoreMotion==7.3
pyobjc-framework-CoreServices==7.3
pyobjc-framework-CoreSpotlight==7.3
pyobjc-framework-CoreText==7.3
pyobjc-framework-CoreWLAN==7.3
pyobjc-framework-CryptoTokenKit==7.3
pyobjc-framework-DeviceCheck==7.3
pyobjc-framework-DictionaryServices==7.3
pyobjc-framework-DiscRecording==7.3
pyobjc-framework-DiscRecordingUI==7.3
pyobjc-framework-DiskArbitration==7.3
pyobjc-framework-DVDPlayback==7.3
pyobjc-framework-EventKit==7.3
pyobjc-framework-ExceptionHandling==7.3
pyobjc-framework-ExecutionPolicy==7.3
pyobjc-framework-ExternalAccessory==7.3
pyobjc-framework-FileProvider==7.3
pyobjc-framework-FileProviderUI==7.3
pyobjc-framework-FinderSync==7.3
pyobjc-framework-FSEvents==7.3
pyobjc-framework-GameCenter==7.3
pyobjc-framework-GameController==7.3
pyobjc-framework-GameKit==7.3
pyobjc-framework-GameplayKit==7.3
pyobjc-framework-ImageCaptureCore==7.3
pyobjc-framework-IMServicePlugIn==7.3
pyobjc-framework-InputMethodKit==7.3
pyobjc-framework-InstallerPlugins==7.3
pyobjc-framework-InstantMessage==7.3
pyobjc-framework-Intents==7.3
pyobjc-framework-IOSurface==7.3
pyobjc-framework-iTunesLibrary==7.3
pyobjc-framework-KernelManagement==7.3
pyobjc-framework-LatentSemanticMapping==7.3
pyobjc-framework-LaunchServices==7.3
pyobjc-framework-libdispatch==7.3
pyobjc-framework-LinkPresentation==7.3
pyobjc-framework-LocalAuthentication==7.3
pyobjc-framework-MapKit==7.3
pyobjc-framework-MediaAccessibility==7.3
pyobjc-framework-MediaLibrary==7.3
pyobjc-framework-MediaPlayer==7.3
pyobjc-framework-MediaToolbox==7.3
pyobjc-framework-Metal==7.3
pyobjc-framework-MetalKit==7.3
pyobjc-framework-MetalPerformanceShaders==7.3
pyobjc-framework-MetalPerformanceShadersGraph==7.3
pyobjc-framework-MLCompute==7.3
pyobjc-framework-ModelIO==7.3
pyobjc-framework-MultipeerConnectivity==7.3
pyobjc-framework-NaturalLanguage==7.3
pyobjc-framework-NetFS==7.3
pyobjc-framework-Network==7.3
pyobjc-framework-NetworkExtension==7.3
pyobjc-framework-NotificationCenter==7.3
pyobjc-framework-OpenDirectory==7.3
pyobjc-framework-OSAKit==7.3
pyobjc-framework-OSLog==7.3
pyobjc-framework-PassKit==7.3
pyobjc-framework-PencilKit==7.3
pyobjc-framework-Photos==7.3
pyobjc-framework-PhotosUI==7.3
pyobjc-framework-PreferencePanes==7.3
pyobjc-framework-PushKit==7.3
pyobjc-framework-Quartz==7.3
pyobjc-framework-QuickLookThumbnailing==7.3
pyobjc-framework-ReplayKit==7.3
pyobjc-framework-SafariServices==7.3
pyobjc-framework-SceneKit==7.3
pyobjc-framework-ScreenSaver==7.3
pyobjc-framework-ScreenTime==7.3
pyobjc-framework-ScriptingBridge==7.3
pyobjc-framework-SearchKit==7.3
pyobjc-framework-Security==7.3
pyobjc-framework-SecurityFoundation==7.3
pyobjc-framework-SecurityInterface==7.3
pyobjc-framework-ServiceManagement==7.3
pyobjc-framework-Social==7.3
pyobjc-framework-SoundAnalysis==7.3
pyobjc-framework-Speech==7.3
pyobjc-framework-SpriteKit==7.3
pyobjc-framework-StoreKit==7.3
pyobjc-framework-SyncServices==7.3
pyobjc-framework-SystemConfiguration==7.3
pyobjc-framework-SystemExtensions==7.3
pyobjc-framework-UniformTypeIdentifiers==7.3
pyobjc-framework-UserNotifications==7.3
pyobjc-framework-UserNotificationsUI==7.3
pyobjc-framework-VideoSubscriberAccount==7.3
pyobjc-framework-VideoToolbox==7.3
pyobjc-framework-Virtualization==7.3
pyobjc-framework-Vision==7.3
pyobjc-framework-WebKit==7.3
pyparsing==3.0.0
PyPDF2==1.26.0
PyPrind==2.11.3
pySmartDL==1.3.4
python-bidi==0.4.2
python-dateutil==2.8.1
python-telegram-bot==13.11
python-telegram-bot-calendar==1.0.5
python-youtube==0.8.1
python3-openid==3.2.0
pyttsx3==2.7
pytube==11.0.1
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rapidfuzz==1.8.2
regex==2022.3.15
reportlab==3.6.6
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==10.13.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
social-auth-app-django==3.1.0
social-auth-core==4.1.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
spotdl==3.9.0
spotipy==2.19.0
sqlparse==0.4.2
tablib==3.1.0
telebot==0.0.4
termcolor==1.1.0
tk==0.1.0
toml==0.10.2
tomlkit==0.9.2
tornado==6.1
tqdm==4.62.3
typing-extensions==3.10.0.2
typing-inspect==0.7.1
tzdata==2021.4
tzlocal==4.0.1
Unidecode==1.3.2
urllib3==1.26.7
uWSGI==2.0.18
validus==0.3.0
virtualenv==20.4.7
webencodings==0.5.1
websockets==10.0
Werkzeug==2.0.2
wget==3.2
wikipedia==1.4.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
youtube-dl==2021.6.6
youtube-search==2.1.0
yt-dlp==2021.11.10.1
ytmusicapi==0.19.4
zipp==3.7.0
